Sourcing test_latest_build.bash automatically scans for the MAC addr
It was confusing to remember all that garbage about having to run the
"scanmacaddrs" command followed by the "findmac" command to get the "right"
MAC address - so the script just goes out and does it for you when you
source it.

Also added the add 'sudo' where needed (permission issues on the factory
laptop).